Make separate file with OperatorFunctions

+++ list04-templates/main.cpp
@@ -1,7 +1,7 @@
 #include "ConsoleController.h"
 
 int main() {
-    ConsoleController<int> consoleController;
+    ConsoleController<double> consoleController;
     consoleController.start();
     return 0;
 }
